requirements.txt
imgaug
aspose-words
cairosvg
bs4
keras==2.11.0
tensorflow==2.11.0
cython
pandas
keras
tqdm
requests
pandas
sklearn
scikit-learn
Pillow
opencv-python
